+++ egui/src/memory.rs
@@ -329,10 +329,9 @@ impl Memory {
         self.areas.layer_id_at(pos, resize_interact_radius_side)
     }
 
-    /// The overall top-most layer. When an area is clicked on or interacted
-    /// with, it is moved above all other areas.
-    pub fn top_most_layer(&self) -> Option<LayerId> {
-        self.areas.order().last().copied()
+    /// An iterator over all layers. Back-to-front. Top is last.
+    pub fn layer_ids(&self) -> impl Iterator<Item = LayerId> + '_ {
+        self.areas.order().iter().copied()
     }
 
     pub(crate) fn had_focus_last_frame(&self, id: Id) -> bool {